website-reinterpretation/requirements.txt
anthropic>=0.18.0
google-genai>=0.2.0
aiohttp>=3.9.0
requests>=2.31.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
playwright>=1.40.0
brotli>=1.1.0
numpy>=1.24.0